Inserido Exceptions

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProdutoService.java
@@ -1,6 +1,8 @@
 package com.example.meu_primeiro_springboot.service;
 
 import org.springframework.stereotype.Service;
+
+import com.example.meu_primeiro_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.meu_primeiro_springboot.model.Produto;
 import com.example.meu_primeiro_springboot.repository.ProdutoRepository;
 import java.util.List;
@@ -19,8 +21,9 @@ public class ProdutoService{
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> BuscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID "+id+" não encontrado."));
     }
 
     public Produto salvarProduto(Produto produto){
@@ -28,6 +31,12 @@ public class ProdutoService{
     }
 
     public void deletarProduto(Long id){
+
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com ID" +id+ "não encontrado.");
+        }
+
+
         produtoRepository.deleteById(id);
     }
 }
